Add files via upload
🚀 Excited to share that I'm gearing up for the upcoming Hackathon! 🚀

I've been diligently preparing by reviewing the Prerequisites and Best Practices for the Hackathon document, which is a goldmine of essential tips and guidelines. If you're participating too, make sure you check it out and get ready to shine! 📄💡

The official details will be released tomorrow at 7:00 AM, and we’ll have 24 hours to put our skills to the test. ⏰ Let's bring our A-game and make this Hackathon a success!

Looking forward to seeing all the innovative solutions and ideas. Best of luck to everyone involved! 💪💻

#Hackathon2024 #Innovation #TechCommunity #ReadyToCode #ProfessionalDevelopment